requirements.txt.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
requests
reportlab
pillow
torch
torchvision